Create pull_request_template.md
In the "Commit message" field, type a short, meaningful commit message that describes the change you made to the file. You can attribute the commit to more than one author in the commit message. 
For more information, see "Creating a commit with multiple authors."  --->
{https://docs.github.com/en/pull-requests/committing-changes-to-your-project/creating-and-editing-commits/creating-a-commit-with-multiple-authors}

Below the commit message fields, decide whether to add your commit to the current branch or to a new branch. If your current branch is the default branch, you should choose to create a new branch for your commit and then create a pull request.
For more information, see "Creating a pull request." --->
{https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/creating-a-pull-request}

**TEMPLATES ARE AVAILABLE TO COLLABORATIONS WHEN THEY ARE MERGED INTO THE REPOSITORY'S DEFAULT BRANCH.**